v3/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Abhijit Gupta <avgupta456@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
fastapi = "^0.101.1"
python-dotenv = "^1.0.0"
uvicorn = "^0.23.2"
gunicorn = "^21.2.0"
requests = "^2.31.0"
SQLAlchemy = "^2.0.20"
sqlalchemy-cockroachdb = "^2.0.1"
psycopg2 = "^2.9.7"
attrs = "^23.1.0"
scipy = "^1.11.1"
CacheControl = "^0.13.1"
typer = {extras = ["all"], version = "^0.9.0"}
black = "^23.7.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
pyright = "^1.1.324"

[tool.poetry.group.dev.dependencies]
pyinstrument = "^4.6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
